requirements.txt
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
PyJWT==2.0.0